+++ src/App.tsx
@@ -1,9 +1,17 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { Outlet } from "react-router-dom";
 import { createGlobalStyle } from "styled-components";
+import {
+  doc,
+  query,
+  collection,
+  where,
+  onSnapshot,
+  getFirestore,
+} from "firebase/firestore";
 import { auth } from "../src/utils/firebase";
 import { onAuthStateChanged } from "firebase/auth";
-import { useDispatch } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import { setUserData } from "../src/actions/index";
 import firebaseapi from "../src/utils/firebaseapi";
 
@@ -11,6 +19,7 @@ import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Slider from "./components/Slider";
 import ScrollToTop from "./components/ScrollToTop";
+import Notification from "./components/modal/Notification";
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -35,7 +44,11 @@ a {
 `;
 
 function App() {
+  const db = getFirestore();
+  const [showNotification, setShowNotification] = useState(false);
   const dispatch = useDispatch();
+  const [getInvitationList, setGetInvitationList] = useState<any>();
+  const [arrayLength, setArrayLength] = useState(0);
 
   useEffect(() => {
     onAuthStateChanged(auth, (user) => {
@@ -50,17 +63,38 @@ function App() {
                 result["main_photo"]
               )
             );
+            // see friend requests difference
+            getFriend(result["user_id"]);
+            //
+            // see chatroom update
+            //
           }
         });
       }
     });
   }, []);
 
+  const getFriend = (id: string) => {
+    onSnapshot(doc(collection(db, "Users"), id), (doc) => {
+      if (doc.exists()) {
+        setGetInvitationList(doc.data().friend_request);
+        // console.log(doc.data().friend_request);
+        // To compare the friend request
+        setArrayLength(doc.data().friend_request.length);
+        setShowNotification(false);
+        if (doc.data().friend_request.length > arrayLength) {
+          setShowNotification(true);
+        }
+      }
+    });
+  };
+
   return (
     <>
       <GlobalStyle />
       <Header />
       <Outlet />
+      {showNotification && <Notification />}
       <Slider />
       <ScrollToTop />
       <Footer />
